requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-WTF==1.2.2
WTForms==3.2.1
python-dotenv==1.1.1
email-validator==2.3.0
boto3>=1.34.0
gunicorn==22.0.0
